+++ src/Recyclarr/Logging/LoggerFactory.cs
@@ -1,6 +1,8 @@
 using System.IO.Abstractions;
 using Serilog;
 using Serilog.Events;
+using Serilog.Templates;
+using Serilog.Templates.Themes;
 using TrashLib.Startup;
 
 namespace Recyclarr.Logging;
@@ -9,11 +11,32 @@ public class LoggerFactory
 {
     private readonly IAppPaths _paths;
 
-    private const string ConsoleTemplate = "[{Level:u3}] {Message:lj}{NewLine}{Exception}";
+    private const string BaseTemplate =
+        "{@m}" +
+        "{@x}" +
+        "\n";
 
     public LoggerFactory(IAppPaths paths)
     {
         _paths = paths;
+    }
+
+    private static ExpressionTemplate GetConsoleTemplate()
+    {
+        const string template =
+            "[{@l:u3}] " +
+            BaseTemplate;
+
+        return new ExpressionTemplate(template, theme: TemplateTheme.Code);
+    }
+
+    private static ExpressionTemplate GetFileTemplate()
+    {
+        const string template =
+            "[{@t:HH:mm:ss} {@l:u3}] " +
+            BaseTemplate;
+
+        return new ExpressionTemplate(template);
     }
 
     public ILogger Create(LogEventLevel level)
@@ -22,8 +45,8 @@ public class LoggerFactory
 
         return new LoggerConfiguration()
             .MinimumLevel.Is(LogEventLevel.Debug)
-            .WriteTo.Console(outputTemplate: ConsoleTemplate, restrictedToMinimumLevel: level)
-            .WriteTo.File(logPath.FullName)
+            .WriteTo.Console(GetConsoleTemplate(), level)
+            .WriteTo.File(GetFileTemplate(), logPath.FullName)
             .CreateLogger();
     }
 }