+++ src/TrashLib/Repo/VersionControl/GitRepositoryFactory.cs
@@ -1,43 +1,18 @@
-using System.IO.Abstractions;
-using System.Text;
-using CliWrap;
-using Common;
 using Serilog;
 
 namespace TrashLib.Repo.VersionControl;
 
 public class GitRepositoryFactory : IGitRepositoryFactory
 {
-    private readonly IFileUtilities _fileUtils;
     private readonly Func<string, IGitRepository> _repoFactory;
-    private readonly IGitPath _gitPath;
     private readonly ILogger _log;
 
     public GitRepositoryFactory(
-        IFileUtilities fileUtils,
         Func<string, IGitRepository> repoFactory,
-        IGitPath gitPath,
         ILogger log)
     {
-        _fileUtils = fileUtils;
         _repoFactory = repoFactory;
-        _gitPath = gitPath;
         _log = log;
-    }
-
-    private async Task<(bool Succeeded, string OutputMsg, string ErrorMsg)> IsValid(IDirectoryInfo repoPath)
-    {
-        var output = new StringBuilder();
-        var error = new StringBuilder();
-        var result = await Cli.Wrap(_gitPath.Path)
-            .WithArguments("status")
-            .WithStandardErrorPipe(PipeTarget.ToStringBuilder(error))
-            .WithStandardOutputPipe(PipeTarget.ToStringBuilder(output))
-            .WithValidation(CommandResultValidation.None)
-            .WithWorkingDirectory(repoPath.FullName)
-            .ExecuteAsync();
-
-        return (result.ExitCode == 0, output.ToString(), error.ToString());
     }
 
     public async Task<IGitRepository> CreateAndCloneIfNeeded(string repoUrl, string repoPath, string branch)
@@ -51,16 +26,9 @@ public class GitRepositoryFactory : IGitRepositoryFactory
         }
         else
         {
-            var validResult = await IsValid(repo.Path);
-            if (!validResult.Succeeded)
-            {
-                _log.Information("Git repository is invalid; deleting and cloning again");
-                _log.Debug("Validity Check Output: {Message}", validResult.OutputMsg.Trim());
-                _log.Debug("Validity Check Error: {Message}", validResult.ErrorMsg.Trim());
-
-                _fileUtils.DeleteReadOnlyDirectory(repo.Path.FullName);
-                await repo.Clone(repoUrl, branch);
-            }
+            // Run just to check repository health. If unhealthy, an exception will
+            // be thrown. That exception will propagate up and result in a re-clone.
+            await repo.Status();
         }
 
         _log.Debug("Remote 'origin' set to {Url}", repoUrl);