+++ src/TrashLib/Repo/VersionControl/IGitRepository.cs
@@ -10,4 +10,5 @@ public interface IGitRepository : IDisposable
     Task SetRemote(string name, string newUrl);
     IDirectoryInfo Path { get; }
     Task Clone(string cloneUrl, string? branch = null);
+    Task Status();
 }